Final AutoValue implementation
